+++ src/test/resources/org/sonar/plugins/delphi/symbol/overloads/imports/IntegerFoo.pas
@@ -1,0 +1,16 @@
+unit IntegerFoo;
+
+{This is a sample Delphi file.}
+
+interface
+
+procedure Foo(Bar: Integer); overload;
+
+implementation
+
+procedure Foo(Bar: Integer);
+begin
+  // Do nothing
+end;
+
+end.